renaming

+++ FinalProject/src/by/epamtc/dubovik/shop/dao/CommentDAO.java
@@ -7,9 +7,9 @@ import by.epamtc.dubovik.shop.entity.Comment;
 
 public interface CommentDAO extends GenericIntIdDAO<Comment> {
 
-	public List<Comment> findByProduct(int productId) throws DAOException;
-	public List<Comment> findByUser(int userIds) throws DAOException;
-	public int calculateCommentCount(int productId) throws DAOException;
-	public int calculateAverageRating(int productId) throws DAOException;
+	public List<Comment> findByProduct(long productId) throws DAOException;
+	public List<Comment> findByUser(long userIds) throws DAOException;
+	public int calculateCommentCount(long productId) throws DAOException;
+	public int calculateAverageRating(long productId) throws DAOException;
 	
 }
