+++ FinalProject/src/by/epamtc/dubovik/shop/dao/ProductForCartDAO.java
@@ -4,6 +4,6 @@ import by.epamtc.dubovik.shop.entity.ProductForCart;
 
 public interface ProductForCartDAO {
 	
-	public ProductForCart findById(int id) throws DAOException;
+	public ProductForCart findById(long id) throws DAOException;
 
 }
