+++ FinalProject/src/by/epamtc/dubovik/shop/dao/UserDAO.java
@@ -7,7 +7,7 @@ import by.epamtc.dubovik.shop.entity.UserLogged;
 
 public interface UserDAO extends GenericIntIdDAO<User> {
 	
-	public boolean blacklistUser(int userId) throws DAOException;
+	public boolean blacklistUser(long userId) throws DAOException;
 	public User findUserInfoByLogin(String login) throws DAOException;
 	public UserLogged authorize(UserForLogin user) throws DAOException;
 	public byte[] takePassword(String login) throws DAOException;
