+++ FinalProject/src/by/epamtc/dubovik/shop/dao/generic/GenericIntIdDAO.java
@@ -4,7 +4,7 @@ import by.epamtc.dubovik.shop.dao.DAOException;
 
 public interface GenericIntIdDAO<T> extends GenericDAO<T> {
 	
-	public T findById(int id) throws DAOException;
-	public boolean delete(int id) throws DAOException;
+	public T findById(long id) throws DAOException;
+	public boolean delete(long id) throws DAOException;
 
 }
