+++ FinalProject/src/by/epamtc/dubovik/shop/service/ProductCategoryService.java
@@ -9,6 +9,6 @@ public interface ProductCategoryService {
 
 	public List<ProductCategory> takeAllCategories() throws ServiceException;
 	public ProductCategory takeByName(String name) throws ServiceException;
-	public ProductCategory takeById(int id) throws ServiceException;
+	public ProductCategory takeById(long id) throws ServiceException;
 	public boolean createCategory(ProductCategory category) throws ServiceException;
 }
